corrected UI standerdization for fee discount page

+++ apps/fastapi/app/schemas/fee_discount.py
@@ -1,11 +1,14 @@
+
 from typing import Optional
-from pydantic import BaseModel, Field, validator, field_validator
+from pydantic import BaseModel, Field, validator
 from datetime import datetime
+from decimal import Decimal
+
 
 class FeeDiscountBase(BaseModel):
     discount_name: str = Field(..., max_length=150)
     discount_type: str = Field(..., max_length=20)
-    discount_value: float
+    discount_value: Decimal
     fee_category: Optional[str] = None
     applicable_class: Optional[str] = None
     description: Optional[str] = None
@@ -30,18 +33,20 @@ class FeeDiscountBase(BaseModel):
 class FeeDiscountCreate(FeeDiscountBase):
     pass
 
+
 class FeeDiscountUpdate(FeeDiscountBase):
     discount_name: Optional[str] = None
     discount_type: Optional[str] = None
-    discount_value: Optional[float] = None
+    discount_value: Optional[Decimal] = None
     fee_category: Optional[str] = None
     applicable_class: Optional[str] = None
+
 
 class FeeDiscountResponse(BaseModel):
     id: int
     discount_name: str
     discount_type: str
-    discount_value: float
+    discount_value: Decimal
     fee_category: Optional[str] = None
     applicable_class: Optional[str] = None
     description: Optional[str] = None
